k8s-resources/delivery/requirements.txt
# Flask Framework
Flask==1.0.2
redis-py-cluster
requests
aioredis-cluster
